Fixed Speech Recognition Python
Turns out the interpreter for Python was wrong and a few options inhibited the ability to spawn a new Python process with the actual script launched. Now the program works! The only thing to fix is the dynamic update of TMP_Text using FileWatchers but more optimized and in parallel