Emulate Automatic Silence
If we don't get to have a GUI option, this is an alternative that allows us to capture microphone input until silence is detected. Discord and Code Documentation has the attached research paper that implements this methodology.

I'll have to test this functionality out tomorrow or some other time to get these functionalities to work.