Changed TTS.cs
Fixed a new setting to where we can distinguish between two avatars and play their voices separately. I also added a Coroutine that allows for continuous checking of when a audio source has finished playing and updates the character accordingly.

Will need to test in my Windows machine or anyone can test it!