Fixed TTS.cs
Turns out that the update method had issues constantly checking to see if the sync file updated, which caused audio buffering (and sometimes no audio playback).

The added fix is to append a bool variable that checks if there is already an instance defined. This instance gets erased with the coroutine.